requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.4,<3.13.0
flake8>=3.6.0,<3.7.0
Pillow>=5.3.0,<5.4.0
psycopg2>=2.7.5,<2.8.0